requirements.txt
fastapi
uvicorn
opencv-python
starlette
python-dotenv
itsdangerous